Added grouop chat invites

+++ backend/controller/chatController.js
@@ -4,6 +4,7 @@ const ChatUser = require('../models/chatUser');
 const User = require('../models/user');
 const sequelize = require('../config/dbConfig');
 const { Op } = require('sequelize');
+const Friendship = require('../models/friendship');
 
 exports.createChat = async (req, res) => {
   try {
@@ -18,6 +19,7 @@ exports.createChat = async (req, res) => {
           where: {
             isGroup: false,
           },
+          required: true,
         },
       ],
       group: ['chatId'],
@@ -64,20 +66,30 @@ exports.acceptChatRequest = async (req, res) => {
 
 exports.declineChatRequest = async (req, res) => {
   try {
-    const { chatId } = req.body;
+    const { chatId, userId } = req.body;
+
     await ChatUser.destroy({
       where: {
         chatId: chatId,
+        userId: userId,
       },
     });
 
-    await Chat.destroy({
+    const usersLeft = await ChatUser.findAll({
       where: {
         chatId: chatId,
       },
     });
 
-    res.status(200).send('sucess');
+    if (usersLeft.length > 1) {
+      await Chat.destroy({
+        where: {
+          chatId: chatId,
+        },
+      });
+    }
+
+    res.status(200).send('success');
   } catch (error) {
     res.status(500).send({ error: 'Failed to decline chat request', message: error.message });
   }
@@ -85,11 +97,108 @@ exports.declineChatRequest = async (req, res) => {
 
 exports.sendChatRequest = async (req, res) => {
   try {
-    const { chatId, userId } = req.body;
+    const { chatId, username, userId } = req.body;
+
+    const user = await User.findOne({
+      where: {
+        username: username,
+      },
+    });
+
+    if (!user) {
+      res.status(400).send({ error: 'No user found' });
+      return;
+    }
+
+    const friendship = await Friendship.findOne({
+      where: {
+        [Op.or]: [
+          { user_id_1: userId, user_id_2: user.userId },
+          { user_id_1: user.userId, user_id_2: userId },
+        ],
+      },
+    });
+
+    if (!friendship) {
+      res.status(400).send({ error: 'Not friends with this user' });
+      return;
+    }
+
+    const existingChatUser = await ChatUser.findOne({
+      where: {
+        userId: user.userId,
+        chatId: chatId,
+      },
+    });
+
+    if (existingChatUser) {
+      res.status(400).send({ error: 'User is already in the chat' });
+      return;
+    }
+
+    const allChatUsers = await ChatUser.findAll({
+      where: {
+        chatId: chatId,
+      },
+      include: [
+        {
+          model: User,
+          attributes: ['username'],
+        },
+      ],
+    });
+
+    const allChatUsernames = [...allChatUsers.map((u) => u.User.username), username];
+
+    const existingChatIds = await ChatUser.findAll({
+      group: ['chatId'],
+      include: [
+        {
+          model: User,
+          where: {
+            username: allChatUsernames,
+          },
+        },
+      ],
+      attributes: ['chatId'],
+    });
+
+    for (const chat of existingChatIds) {
+      let allUsers = await ChatUser.findAll({
+        where: {
+          chatId: chat.chatId,
+        },
+        include: [
+          {
+            model: User,
+            attributes: ['username'],
+          },
+        ],
+      });
+
+      if (allUsers.length === allChatUsernames.length) {
+        allUsers = allUsers.map((u) => u.User.username);
+
+        if (allUsers.every((element) => allChatUsernames.includes(element))) {
+          res.status(400).send({ error: 'Group chat already exists' });
+          return;
+        }
+      }
+    }
+
     const chatUser = await ChatUser.create({
       chatId: chatId,
-      userId: userId,
+      userId: user.userId,
     });
+
+    const groupChat = await Chat.findOne({
+      where: {
+        chatId: chatId,
+      },
+    });
+    groupChat.isGroup = true;
+    await groupChat.save();
+
     res.status(200).send(chatUser);
   } catch (error) {
     res.status(500).send({ error: 'Failed to send chat request', message: error.message });
@@ -152,7 +261,6 @@ exports.getChatMessages = async (req, res) => {
         chatId: chatId,
       },
     });
-    console.log(chatMessages);
 
     res.status(200).send(chatMessages);
   } catch (error) {
@@ -163,11 +271,16 @@ exports.getChatMessages = async (req, res) => {
 exports.getChatStatus = async (req, res) => {
   try {
     const { chatId } = req.body;
-    console.log(chatId);
     const chatUsers = await ChatUser.findAll({
       where: {
         chatId: chatId,
       },
+      include: [
+        {
+          model: User,
+          require: true,
+        },
+      ],
     });
     res.status(200).send(chatUsers);
   } catch (error) {
@@ -177,7 +290,6 @@ exports.getChatStatus = async (req, res) => {
 
 exports.sendMessage = async (message, chatId, userId) => {
   try {
-    console.log({ message, chatId, userId });
     const chatMessage = await ChatMessage.create({
       chatId: chatId,
       text: message,
